Remove deprecated components
In v4.x we deprecated many components, this removes them all to clean up
our codebase a bit.
