+++ test/functional/HmacTokenTest.php
@@ -6,8 +6,8 @@ namespace Lcobucci\JWT\FunctionalTests;
 use Lcobucci\JWT\Configuration;
 use Lcobucci\JWT\Signer\Hmac\Sha256;
 use Lcobucci\JWT\Signer\Hmac\Sha512;
-use Lcobucci\JWT\Signer\Hmac\UnsafeSha256;
 use Lcobucci\JWT\Signer\Key\InMemory;
+use Lcobucci\JWT\SodiumBase64Polyfill;
 use Lcobucci\JWT\Token;
 use Lcobucci\JWT\Validation\Constraint\SignedWith;
 use Lcobucci\JWT\Validation\RequiredConstraintsViolated;
@@ -136,11 +136,12 @@ class HmacTokenTest extends TestCase
     public function everythingShouldWorkWhenUsingATokenGeneratedByOtherLibs(): void
     {
         $config = Configuration::forSymmetricSigner(
-            new UnsafeSha256(),
-            InMemory::plainText('testing'),
+            new Sha256(),
+            InMemory::base64Encoded('FkL2+V+1k2auI3xxTz/2skChDQVVjT9PW1/grXafg3M='),
         );
-        $data   = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXUyJ9.eyJoZWxsbyI6IndvcmxkIn0.Rh'
-                . '7AEgqCB7zae1PkgIlvOpeyw9Ab8NGTbeOH7heHO0o';
+
+        $data = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJoZWxsbyI6IndvcmxkIn0.'
+              . 'ZQfnc_iFebE--gXmnhJrqMXv3GWdH9uvdkFXTgBcMFw';
 
         $token = $config->parser()->parse($data);
         assert($token instanceof Token\Plain);
@@ -153,14 +154,20 @@ class HmacTokenTest extends TestCase
     /** @test */
     public function signatureValidationWithLocalFileKeyReferenceWillOperateWithKeyContents(): void
     {
-        $key = tempnam(sys_get_temp_dir(), 'key');
+        $key = tempnam(sys_get_temp_dir(), 'a-very-long-prefix-to-create-a-longer-key');
         assert(is_string($key));
 
-        file_put_contents($key, 'just a dummy key');
+        file_put_contents(
+            $key,
+            SodiumBase64Polyfill::base642bin(
+                'FkL2+V+1k2auI3xxTz/2skChDQVVjT9PW1/grXafg3M=',
+                SodiumBase64Polyfill::SODIUM_BASE64_VARIANT_ORIGINAL,
+            ),
+        );
 
         $validKey      = InMemory::file($key);
         $invalidKey    = InMemory::plainText('file://' . $key);
-        $signer        = new UnsafeSha256();
+        $signer        = new Sha256();
         $configuration = Configuration::forSymmetricSigner($signer, $validKey);
         $validator     = $configuration->validator();
